pyproject.toml
[project]
name = "reporium-audit"
version = "1.0.0"
description = "Nightly automated audit of the entire Reporium platform"
requires-python = ">=3.11"
dependencies = ["httpx>=0.27", "python-dotenv>=1.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21", "ruff>=0.4"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
